Move UnlockedWithDocumentsView to official_docuemnts app
Also fix the view and test. This has been giving a 500 for a while and
isn't linked to anywhere, so I'm not worried about setting up a
redirect.

Closes #270
