+++ ynr/apps/official_documents/templates/official_documents/unlocked_with_documents.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 
 {% block content %}
-<h2>Areas with nomination papers that need importing</h2>
+<h2>Unlocked posts with nomination papers</h2>
 {% for doc in unlocked_sopns %}
   <li>
     <a href="{% url 'bulk_add' election=doc.election.slug post_id=doc.post.extra.slug %}?source={{ doc.source_url }}">
