+++ ynr/apps/official_documents/urls.py
@@ -16,4 +16,7 @@ urlpatterns = [
     url(r'^(?P<pk>\d+)/$',
         views.DocumentView.as_view(),
         name='uploaded_document_view'),
+    url(r'^uploaded/$',
+        views.UnlockedWithDocumentsView.as_view(),
+        name='unlocked_posts_with_documents'),
 ]
